+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# development version
+
 # cli 2.3.1
 
 * ANSI color support detection works correctly now in older RStudio,